Extract common code to users/crandel
