+++ keyboards/crkbd/keymaps/crandel/keymap.c
@@ -17,76 +17,6 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include QMK_KEYBOARD_H
-#include "vial.h"
-
-#include "crd_keycodes.h"
-
-#ifdef OLED_ENABLE
-#  include <stdio.h>
-#  include "crd_oled.c"
-#endif
-
-#include "crd_helpers.c"
-
-#ifdef LEADER_ENABLE
-#  include "crd_leader.c"
-#endif
-
-#ifdef TAP_DANCE_ENABLE
-#  include "crd_tapdance.c"
-#endif
-
-#ifdef RGBLIGHT_ENABLE
-layer_state_t default_layer_state_set_user(layer_state_t state) {
-  #ifdef CONSOLE_ENABLE
-  uprintf("current state in default_layer_state_set_user %d, biton %d\n", state, biton32(state));
-  #endif // CONSOLE_ENABLE
-  switch (biton32(state)) {
-    case COLMAK_L:
-      set_indicators_state(CR_HSV_GREEN, "Colemak");
-      break;
-    case NUM_L:
-      set_indicators_state(CR_HSV_BLUE, "Numbers");
-      break;
-    case CHARS_L:
-      set_indicators_state(CR_HSV_GOLD, "Chars");
-      break;
-    case MOUSE_L:
-      set_indicators_state(CR_HSV_ORANGE, "Mouse");
-      break;
-    case GAME_L:
-      set_indicators_state(CR_HSV_RED, "Games");
-      break;
-  }
-  return state;
-};
-#endif // RGBLIGHT_ENABLE
-
-void set_lang(bool lng){
-  if (lng) {
-    tap_code(KC_CAPS);
-  } else {
-    tap_code16(S(KC_CAPS));
-  }
-}
-
-bool process_record_user(uint16_t keycode, keyrecord_t *record) {
-  switch (keycode) {
-  case V_US:
-    if (record->event.pressed) {
-      set_lang(true);
-    }
-    break;
-  case V_UK:
-    if (record->event.pressed) {
-      set_lang(false);
-    }
-    break;
-  }
-  return true;
-}
-
 void keyboard_post_init_user(void) {
   // Call the post init code.
   #ifdef TAP_DANCE_ENABLE
@@ -110,27 +40,27 @@ uint8_t mod_config(uint8_t mod) {
 
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
   [COLMAK_L] = LAYOUT_split_3x6_3(
-  //,----------------------------------------------------------------------.    ,--------------------------------------------------------------------------.
-     XXXXXXX, LGUI_T(KC_Q), LALT_T(KC_W), LCTL_T(KC_F), LSFT_T(KC_P),  KC_B,       KC_E, LSFT_T(KC_I), LCTL_T(KC_U), RALT_T(KC_Y), LGUI_T(KC_SCLN), XXXXXXX,
-  //|-------+-------------+-------------+-------------+-------------+------|    |------+-------------+-------------+-------------+----------------+--------|
-     XXXXXXX,         KC_A,         KC_R,         KC_S,         KC_T,  KC_G,       KC_M,         KC_H,         KC_J,         KC_K,            KC_L, XXXXXXX,
-  //|-------+-------------+-------------+-------------+-------------+------|    |------+-------------+-------------+-------------+----------------+--------|
-     XXXXXXX,         KC_Z,         KC_X,         KC_C,         KC_D,  KC_V,       KC_O,         KC_N,      KC_COMM,       KC_DOT,          SLS_TD, XXXXXXX,
-  //|-------+-------------+-------------+-------------+-------------+-- ---|    |------+-------------+-------------+-------------+----------------+--------|
-                                               QK_LEAD,       KC_SPC, TD(0),     KC_ENT,       KC_TAB,      KC_BSPC
-                                               //`-------------------------'    `----------------------------------'
+  //,-------------------------------.    ,---------------------------.
+     XXXXXXX, ____COLEMAK_L11____,       ____COLEMAK_L12____, XXXXXXX,
+  //|-------+--------+-------+------|    |------+-------+---+--------|
+     XXXXXXX, ____COLEMAK_L21____,       ____COLEMAK_L22____, XXXXXXX,
+  //|-------+--------+-------+------|    |------+-------+---+--------|
+     XXXXXXX, ____COLEMAK_L31____,       ____COLEMAK_L32____, XXXXXXX,
+  //|-------+--------+-------+------|    |------+-------+---+--------|
+                          __THUMB1__,     __THUMB2__
+                        //`---------'    `----------'
   ),
 
   [NUM_L] = LAYOUT_split_3x6_3(
-  //,--------------------------------------------------------------------------.    ,---------------------------------------------------------------------------------------.
-      XXXXXXX, LGUI_T(KC_F5), LALT_T(KC_7), LCTL_T(KC_4), LSFT_T(KC_2), KC_MINS,          KC_HOME, LSFT_T(KC_END), LCTL_T(KC_PGDN), RALT_T(KC_PGUP), LGUI_T(KC_DOT), XXXXXXX,
-  //|--------+--------------+-------------+-------------+-------------+--------|    |------------+---------------+----------------+----------------+---------------+--------|
-      XXXXXXX,         KC_F4,         KC_8,         KC_5,         KC_1,    KC_3,           KC_DOT,        KC_LEFT,         KC_DOWN,           KC_UP,       KC_RIGHT, XXXXXXX,
-  //|--------+--------------+-------------+-------------+-------------+--------|    |------------+---------------+----------------+----------------+---------------+--------|
-      XXXXXXX,         KC_F3,         KC_9,         KC_6,         KC_0,   KC_F2,          KC_BTN4,        KC_BTN1,         KC_BTN2,         KC_BTN3,        KC_BSLS, XXXXXXX,
-  //---------+--------------+-------------+-------------+-------------+--------|    |------------+---------------+----------------+----------------+---------------+--------|
-                                                 QK_LEAD,       KC_SPC,   TD(0),           KC_ENT,         KC_TAB,         KC_BSPC
-                                                 //`---------------------------'    `---------------------------------------------'
+  //,-------------------------------.    ,---------------------------.
+     XXXXXXX, ____NUMBERS_L11____,       ____NUMBERS_L12____, XXXXXXX,
+  //|-------+--------+-------+------|    |------+-------+---+--------|
+     XXXXXXX, ____NUMBERS_L21____,       ____NUMBERS_L22____, XXXXXXX,
+  //|-------+--------+-------+------|    |------+-------+---+--------|
+     XXXXXXX, ____NUMBERS_L31____,       ____NUMBERS_L32____, XXXXXXX,
+  //|-------+--------+-------+------|    |------+-------+---+--------|
+                          __THUMB1__,     __THUMB2__
+                        //`---------'    `----------'
   ),
 
   [CHARS_L] = LAYOUT_split_3x6_3(
