+++ keyboards/crkbd/keymaps/crandel/rules.mk
@@ -21,3 +21,4 @@ VIAL_ENABLE            = yes # Enable VIAL
 VIAL_INSECURE          = yes # Disable security
 VIA_ENABLE             = yes # Enable VIA
 VIRTSER_ENABLE         = no
+SRC += crandel.c
