+++ users/crandel/crd_helpers.c
