+++ users/crandel/crd_leader.c
